scraper/requirements.txt
selenium>=4.15.0
undetected-chromedriver>=3.5.0
beautifulsoup4>=4.12.0
gspread>=5.12.0
google-auth>=2.23.0
webdriver-manager>=4.0.0
requests>=2.31.0
Pillow>=10.0.0
supabase>=2.0.0
python-dotenv>=1.0.0
